Create and use mock data for posts
